backend/requirements.txt
Flask>=3.0.3
bcrypt>=4.0.1
PyJWT>=2.8.0
Werkzeug>=3.0.2
psycopg2-binary>=2.9.9
redis>=5.0.4
celery>=5.4.0
Flask-SQLAlchemy>=3.1.1
Flask-Migrate>=4.0.7
geopy>=2.4.1
spacy>=3.7.4
marshmallow>=3.21.1
python-dotenv>=1.0.1
structlog>=24.1.0
prometheus-flask-exporter>=0.23.0
kafka-python-ng>=2.2.2
streamlit>=1.32.0
pytest>=8.1.1
pytest-flask>=1.3.0
feedparser>=6.0.11
aiohttp>=3.9.3
Flask-SocketIO>=5.3.6
eventlet>=0.35.2
Flask-Cors>=4.0.0
GeoAlchemy2>=0.14.3
scikit-learn>=1.4.1
xgboost>=2.0.3
sentinelsat>=1.2.1
rasterio>=1.3.9
geopandas>=0.14.3
numpy>=1.26.4
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
